+++ tests/integ/modin/frame/test_iat.py
@@ -290,3 +290,12 @@ def test_iat_neg(
 ):
     with pytest.raises(error):
         default_index_snowpark_pandas_df.iat[key]
+
+
+@sql_count_checker(query_count=0)
+def test_raise_set_cell_with_list_like_value_error():
+    s = pd.Series([[1, 2], [3, 4]])
+    with pytest.raises(NotImplementedError):
+        s.iat[0] = [0, 0]
+    with pytest.raises(NotImplementedError):
+        s.to_frame().iat[0, 0] = [0, 0]